+++ src/reducers/tx_count_by_native_token_policy_id.rs
@@ -65,10 +65,10 @@ impl Reducer {
                             let number_of_minted_or_destroyed = assets.len();
 
                             let key = self.config_key(policy_id, epoch_no);
-                            
+
                             let crdt = model::CRDTCommand::PNCounter(
                                 key,
-                                number_of_minted_or_destroyed as i64,
+                                number_of_minted_or_destroyed as i128,
                             );
                             output.send(gasket::messaging::Message::from(crdt))?;
                         }
@@ -85,7 +85,7 @@ impl Config {
     pub fn plugin(self,
         chain: &crosscut::ChainWellKnownInfo
     ) -> super::Reducer {
-        let reducer = Reducer { 
+        let reducer = Reducer {
             config: self,
             chain: chain.clone(),
          };
